backend/requirements.txt
anaconda-client==1.10.0
anaconda-navigator==2.2.0
autopep8==1.5.7
backports.weakref==1.0.post1
brotlipy==0.7.0
clyent==1.2.2
conda==22.9.0
conda-build==3.22.0
conda-verify==3.4.2
contourpy==1.0.5
cssbeautifier==1.14.7
cycler==0.11.0
distlib==0.3.6
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djlint==1.19.3
EditorConfig==0.12.3
fonttools==4.37.4
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==5.0.0
jsbeautifier==1.14.7
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.6.0
navigator-updater==0.3.0
numpy==1.23.3
pandas==1.5.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.6.2
ply==3.11
pycodestyle==2.7.0
pycosat==0.6.3
PyQt5==5.15.7
pywin32==302
PyYAML==6.0
regex==2022.10.31
toml==0.10.2
tomli==2.0.1
virtualenv==20.17.1
webencodings==0.5.1
wincertstore==0.2
zipp==3.10.0
